# api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/loans/return":{"put":{"tags":["loan-controller"],"operationId":"returnBook","parameters":[{"name":"bookId","in":"query","required":true,"schema":{"type":"string"}},{"name":"userId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/books/{id}":{"get":{"tags":["book-controller"],"operationId":"getBookById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Book"}}}}}},"put":{"tags":["book-controller"],"operationId":"updateBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"title","in":"query","required":true,"schema":{"type":"string"}},{"name":"autor","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["book-controller"],"operationId":"deleteBookId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"addUser","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/loans":{"get":{"tags":["loan-controller"],"operationId":"getActiveLoans","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Loan"}}}}}}},"post":{"tags":["loan-controller"],"operationId":"loanBook","parameters":[{"name":"bookId","in":"query","required":true,"schema":{"type":"string"}},{"name":"userId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Loan"}}}}}}},"/api/books":{"get":{"tags":["book-controller"],"operationId":"getAllBooks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}},"post":{"tags":["book-controller"],"operationId":"addBook","parameters":[{"name":"title","in":"query","required":true,"schema":{"type":"string"}},{"name":"autor","in":"query","required":true,"schema":{"type":"string"}},{"name":"id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/loans/available":{"get":{"tags":["loan-controller"],"operationId":"isBookAvailable","parameters":[{"name":"bookId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/books/exists/{id}":{"get":{"tags":["book-controller"],"operationId":"existsBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/books/autor/{autor}":{"get":{"tags":["book-controller"],"operationId":"getBooksByAutor","parameters":[{"name":"autor","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}}}},"components":{"schemas":{"Book":{"type":"object","properties":{"title":{"type":"string"},"autor":{"type":"string"},"id":{"type":"string"}}},"Loan":{"type":"object","properties":{"book":{"$ref":"#/components/schemas/Book"},"user":{"$ref":"#/components/schemas/User"},"loanDate":{"type":"string","format":"date"},"status":{"type":"string","enum":["ACTIVE","RETURNED"]}}},"User":{"type":"object","properties":{"name":{"type":"string"},"id":{"type":"string"}}}}}}